server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = "Writing the tabula rasa server"
authors = ["Sebastian R. Valdivia <75336522+SebastianRValdivia@users.noreply.github.com>"]
license = "GPL"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.6"
django-rest-framework = "^0.1.0"
django-cors-headers = "^3.13.0"
django-environ = "^0.9.0"
django-filter = "^22.1"
Pillow = "^9.3.0"
selenium = "^4.7.2"
Markdown = "^3.4.1"
django-extensions = "^3.2.1"
Werkzeug = "^2.2.2"
django-constance = "^2.9.1"
django-picklefield = "^3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
